refactor: 홈 패키지 구조 정리
