+++ feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/constants/GuideStep.kt
@@ -5,7 +5,7 @@
  * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
  */
 
-package team.duckie.app.android.feature.ui.home.screen.guide
+package team.duckie.app.android.feature.ui.home.constants
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
