+++ feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/screen/guide/HomeGuideFeatureScrren.kt
@@ -31,6 +31,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import team.duckie.app.android.feature.ui.home.R
 import team.duckie.app.android.feature.ui.home.component.BaseBottomLayout
+import team.duckie.app.android.feature.ui.home.constants.GuideStep
 import team.duckie.app.android.shared.ui.compose.DuckieHorizontalPagerIndicator
 import team.duckie.quackquack.ui.color.QuackColor
 import team.duckie.quackquack.ui.component.QuackBody2
