+++ feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/screen/home/HomeScreen.kt
@@ -5,7 +5,7 @@
  * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
  */
 
-package team.duckie.app.android.feature.ui.home.screen
+package team.duckie.app.android.feature.ui.home.screen.home
 
 import androidx.compose.animation.Crossfade
 import androidx.compose.foundation.layout.Box
@@ -17,13 +17,13 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import org.orbitmvi.orbit.compose.collectAsState
 import team.duckie.app.android.feature.ui.home.constants.HomeStep
-import team.duckie.app.android.feature.ui.home.viewmodel.HomeViewModel
+import team.duckie.app.android.feature.ui.home.viewmodel.home.HomeViewModel
 import team.duckie.app.android.util.compose.activityViewModel
 
 private val HomeHorizontalPadding = PaddingValues(horizontal = 16.dp)
 
 @Composable
-internal fun DuckieHomeScreen(
+internal fun HomeScreen(
     vm: HomeViewModel = activityViewModel(),
 ) {
     val state = vm.collectAsState().value
