+++ feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/viewmodel/guide/HomeGuideState.kt
@@ -5,6 +5,8 @@
  * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
  */
 
-package team.duckie.app.android.feature.ui.home.screen.guide.viewmodel
+package team.duckie.app.android.feature.ui.home.viewmodel.guide
 
-sealed class HomeGuideSideEffect
+internal data class HomeGuideState(
+    val isGuideStarted: Boolean = true,
+)
