+++ feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/viewmodel/home/HomeSideEffect.kt
@@ -5,7 +5,7 @@
  * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
  */
 
-package team.duckie.app.android.feature.ui.home.viewmodel.sideeffect
+package team.duckie.app.android.feature.ui.home.viewmodel.home
 
 internal sealed class HomeSideEffect {
