+++ feature-ui-home/src/main/kotlin/team/duckie/app/android/feature/ui/home/viewmodel/ranking/RankingSideEffect.kt
@@ -5,14 +5,7 @@
  * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
  */
 
-/*
- * Designed and developed by Duckie Team, 2022
- *
- * Licensed under the MIT.
- * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
- */
-
-package team.duckie.app.android.feature.ui.home.screen.ranking.sideeffect
+package team.duckie.app.android.feature.ui.home.viewmodel.ranking
 
 internal sealed class RankingSideEffect {
     class ReportError(val exception: Throwable) : RankingSideEffect()
